+++ src/main/MainB.js
@@ -1,6 +1,6 @@
-import "welfare/main/Main.css";
-import HeaderB from "welfare/main/component/header/HeaderB";
-import ButtonGroup from "welfare/main/component/serviceBtn/ButtonGroup";
+import HeaderB from "main/component/header/HeaderB";
+import ButtonGroup from "main/component/serviceBtn/ButtonGroup";
+import "main/Main.css";
 import CardTag from "./component/card/CardTag";
 import IsCardSection from "./component/card/IsCardSection";
 import AlarmSection from "./component/payAlarm/AlarmSection";
