+++ src/main/component/header/HeaderB.js
@@ -1,6 +1,6 @@
-import 'welfare/main/component/header/HeaderB.css';
-import LogoNav from "welfare/main/component/header/LogoNav";
-import ProfileB from "welfare/main/component/header/ProfileB";
+import 'main/component/header/HeaderB.css';
+import LogoNav from "main/component/header/LogoNav";
+import ProfileB from "main/component/header/ProfileB";
 
 function Header(props) {
   return (
